Separate container from component as per https://github.com/MetaMask/metamask-extension/pull/12019#discussion_r703804969
