+++ ui/components/app/confirm-page-container/index.js
@@ -1,4 +1,4 @@
-export { default } from './confirm-page-container.component';
+export { default } from './confirm-page-container.container';
 export { default as ConfirmPageContainerHeader } from './confirm-page-container-header';
 export { default as ConfirmDetailRow } from './confirm-detail-row';
 export { default as ConfirmPageContainerNavigation } from './confirm-page-container-navigation';
